# swagger.json
{
  "basePath": "/",
...
  "host": "apijhmhportail-akcc2f0.ew.gateway.dev",
  "info": {
    "description": "powered by Flasgger",
    "termsOfService": "/tos",
    "title": "API Portail JHMH",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/reservations/{confirmation_code}": {
      "get": {
        "description": "Retourne un objet JSON complet contenant toutes les données disponibles pour une réservation,\naprès avoir appliqué une logique de fusion intelligente : pour chaque champ, la valeur `_trace` \n(source) est prioritaire sur la valeur `_value` (calculée).\n",
        "operationId": "get_reservation_by_code",
        "parameters": [
          {
            "description": "Code de confirmation de la réservation.",
            "in": "path",
            "name": "confirmation_code",
            "required": true,
            "type": "string"
          },
          {
            "default": false,
            "description": "Inclure la liste complète des `reservation_logs` dans la réponse.",
            "in": "query",
            "name": "include_logs",
            "required": false,
            "type": "boolean"
          },
          {
            "default": false,
            "description": "Inclure le champ `auditNote` (note d'audit du calcul) dans la réponse.",
            "in": "query",
            "name": "include_audit_note",
            "required": false,
            "type": "boolean"
          },
          {
            "default": false,
            "description": "Force l'inclusion de tous les champs `_trace` et désactive la fusion par défaut.",
            "in": "query",
            "name": "force_trace",
            "required": false,
            "type": "boolean"
          },
          {
            "default": false,
            "description": "Force l'inclusion de tous les champs `_value` et désactive la fusion par défaut. Combiné avec `force_trace`, retourne les données brutes.",
            "in": "query",
            "name": "force_value",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Réservation trouvée avec succès.",
            "schema": {
              "description": "Un objet JSON contenant toutes les paires clé/valeur de la réservation traitée.",
              "type": "object"
            }
          },
          "404": {
            "description": "Réservation non trouvée."
          },
          "500": {
            "description": "Erreur interne du serveur."
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Récupère les détails complets et traités d'une réservation.",
        "tags": ["Reservations"]
      }
    },
...
  },
  "schemes": ["https"],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "description": "Clé API pour l'accès à l'API Gateway",
      "in": "header",
      "name": "x-api-key",
      "type": "apiKey"
    }
  },
  "swagger": "2.0",
...
}